♻️ [#196] 리스너 실행 EventListener로 변경
